Install from CodeArtifact (#13)
## Why is this change necessary?
Need to have grandchild repos be able to install dependencies from AWS
CodeArtifact


 ## How does this change address the issue?
Updated the create-aws-profile.sh script to include some more things by
default, and made it Jinja so that it was more configurable.

Added a script to authenticate into CodeArtifact. it's still a bit
annoying because if you want the environmental variables for uv to
persist in your environment (so you could do a `uv lock` on your own)
you have to run it via source...not sure the best way to work around
that....uv doesn't seem to have a command to update a config like poetry
did



 ## What side effects does this change have?
None


 ## How is this change tested?
In many downstream repos. python-template and aws-central infra child
repos. Grandchild repos: cloud courier, ephemeral pulumi deploy, and a
private repo that installed from codeartifact. CI and local development
was tested.


 ## Other
Made the devcontainer startup not fail if there wasn't a uv lock file.
This should make it easier to instantiate new repos (in case you forget
to create the lock file initially).